fix: 주문 생성 로직 수정 - PG 결재 로직 Facade -> Service 이동 - 쿠폰 계산 BigDecimal -> Money 변경 - 주문 결재 상태 변경, 포인트 적립 기능 추가

+++ apps/commerce-api/src/main/java/com/loopers/domain/order/Money.java
@@ -44,7 +44,7 @@ public class Money {
     return new Money(this.amount.subtract(other.amount), this.currency);
   }
 
-  public Money multiply(long multiplier) {
+  public Money multiply(double multiplier) {
     BigDecimal multiplierDecimal = BigDecimal.valueOf(multiplier);
     BigDecimal newAmount = this.amount.multiply(multiplierDecimal);
     return new Money(newAmount, this.currency);
